Move remaining controllers into Web namespace/group.
